+++ dune/istl/multitypeblockmatrix.hh
@@ -66,17 +66,6 @@ namespace Dune {
       return 1+sizeof...(Args);
     }
 
-    /** \brief Return the number of matrix rows
-     *
-     * \deprecated Use method <code>N</code> instead.
-     *             This will be removed after Dune 2.8.
-     */
-    [[deprecated("Use method 'N' instead")]]
-    static constexpr size_type size()
-    {
-      return 1+sizeof...(Args);
-    }
-
     /** \brief Return the number of matrix columns */
     static constexpr size_type M()
     {