+++ CHANGELOG.md
@@ -17,6 +17,8 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 
 - Removed deprecated function `initSolverFactories()`, use `initSolverFactories<O>` instead.
 
+- Removed deprecated `MultyTypeBlock*::size()`, use `N()` instead.
+
 
 # Release 2.9